requirements.txt
requests==2.26.0
python-dotenv==0.19.1
fastapi==0.70.0
uvicorn==0.15.0
aiohttp==3.8.1
tqdm==4.63.0